requirements.txt
python-dotenv
ShopifyAPI
clickhouse-driver
Flask